BackEnd/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
asgiref==3.8.1
autobahn==24.4.2
twisted[tls]==24.11.0
cryptography==44.0.2
hyperlink==21.0.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1